requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.30.1
opencv-python==4.10.0.82
orjson==3.11.0
packaging==25.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==11.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.19.2
pypdfium2==4.30.0
pytesseract==0.3.10
python-docx==1.1.0
python-dotenv==1.1.1
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.2
rich==14.0.0
rich-toolkit==0.14.8
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.1
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
numpy>=1.26.0,<2
opencv-python
sqlalchemy
psycopg2-binary
passlib[bcrypt]